requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-multipart
anthropic
htmx
sqlalchemy
python-dotenv 